02_basic_retrieval/requirements.txt
langchain>=0.1.2
langchain-qdrant>=0.2.0
langchain-openai>=0.1.1
qdrant-client>=1.6.0
openai>=1.0.0
python-dotenv>=1.0.0
